requirements.txt
pandas
requests
google-genai
openai
python-dotenv
